Solucion error undefined

+++ src/app/components/btn/btn.component.ts
@@ -11,18 +11,20 @@ import { UsersService } from '../../services/users.service';
 })
 export class BtnComponent {
   @Input() container:string = "";
-  @Input() idUser:string = "";
+  @Input() idUser:string | undefined = "";
   usersService = inject(UsersService)
   
 
-  async borrarUser(id: string){
+  async borrarUser(id: string | undefined){
     id = this.idUser 
-    let confirmacion = confirm('¿Deseas eliminar al usuario con id ' + this.idUser)
-    if(confirmacion){
-      let response = await this.usersService.delete(id);
-      if(response._id){
-        alert('Se ha borrado el usuario ' + response.first_name + ' ' + response.last_name + 'con éxito')
-      }
+    if(id !== undefined){
+      let confirmacion = confirm('¿Deseas eliminar al usuario con id ' + this.idUser)
+      if(confirmacion){
+        let response = await this.usersService.delete(id);
+        if(response._id){
+          alert('Se ha borrado el usuario ' + response.first_name + ' ' + response.last_name + 'con éxito')
+        }
+    }
     }
   }
